Creates median and mean abundance barplots, and boxplots for microshades taxa contribution plots keeping colors consistent in the microshades stacked barplots

Creates median and mean abundance barplots, and boxplots for microshades taxa contribution plots keeping colors consistent in the microshades stacked barplots

plot_contributions(mdf, cdf, col_name, data_match, short_name = NULL)

Arguments

mdf

melted data frame with microshades group processing

cdf

color data frame

col_name

name of column to subset from

data_match

specific data to subset match

short_name

alternate group name in both the mdf and cdf, note the group column still must exist

Value

list

"box" boxplot

"mean" barplot of mean abundance

"median"barplot of median abundance

Examples

plots <- plot_contributions(mdf, cdf,  "dmm", 1, short_name = "group_short")
#> Error in eval(expr, envir, enclos): object 'mdf' not found

plots$box
#> Error in eval(expr, envir, enclos): object 'plots' not found
plots$mean
#> Error in eval(expr, envir, enclos): object 'plots' not found
plots$median
#> Error in eval(expr, envir, enclos): object 'plots' not found

plots <- plot_contributions(mdf, cdf,  "sex", "female")
#> Error in eval(expr, envir, enclos): object 'mdf' not found